creato il progetto con intelliJ IDEA

+++ La_Disfida/src/main/java/Finance.java
@@ -1,11 +1,11 @@
 public class Finance {
-    public double calculateLoanPayment(double principal, double annualRate, int years) {
+    public double calculateLoanPayment(double principal, double annualRate, double years) {
         if (principal < 0 || annualRate < 0 || years < 0) {
             throw new IllegalArgumentException("Negative values are not allowed.");
         }
 
         double monthlyRate = annualRate / 100 / 12;
-        int numPayments = years * 12;
+        int numPayments = (int)(years * 12);
 
         if (numPayments == 0) {
             throw new IllegalArgumentException("Loan duration cannot be zero.");
